requirements.txt
numpy==1.24.3
opencv-python==4.10.0.84
torch==2.3.1
torchvision==0.18.1
matplotlib==3.9.0
seaborn==0.13.2
scikit-learn==1.5.0
timm==1.0.3
einops==0.7.0
fvcore==0.1.5.post20221221
pytorch-grad-cam==1.4.0
Pillow==10.3.0
torchinfo==1.8.0
pandas==2.2.2
